tool_recommender_service/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
langchain
